pyproject.toml
[tool.poetry]
name = "ai_assistant"
version = "0.1.0"
description = ""
authors = ["Hrebenozhko Ruslan <hrebenozhko@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
langchain = "^0.3.9"
langchain-openai = "^0.2.10"
langchain-milvus = "^0.1.7"
pyyaml = "^6.0.2"
langchain-community = "^0.3.8"
streamlit = "^1.40.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
insert_docs = "script.insert_docs:main"

